+++ projects/games/battle/src/random.c
@@ -1,7 +1,8 @@
 #include <stdlib.h>
 #include "random.h"
 
-int16_t random_int(int16_t min, int16_t max)
+// Generates a random integer between min and max (inclusive)
+int32_t random_int(int32_t min, int32_t max)
 {
     return min + rand() % (max - min + 1);
 }
